fixing factory functions
