+++ docs/source/vocab.rst
@@ -19,16 +19,6 @@ torchtext.vocab
 
 .. autofunction:: vocab
 
-:hidden:`load_vocab_from_file`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: load_vocab_from_file
-
-:hidden:`build_vocab_from_text_file`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: build_vocab_from_text_file
-
 :hidden:`build_vocab_from_iterator`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
