+++ torchtext/experimental/vocab_factory.py
@@ -1,32 +1,68 @@
+import torch
 from torchtext.vocab import Vocab
+from typing import Optional, Callable
 from torchtext._torchtext import (
-    _build_vocab_from_text_file
+    _build_vocab_from_text_file,
+    _load_vocab_from_file,
+    _build_vocab_from_text_file_using_python_tokenizer,
 )
 
+__all__ = [
+    'build_vocab_from_text_file',
+    'load_vocab_from_file',
+]
 
-def build_vocab_from_text_file(file_path, jited_tokenizer, min_freq=1, num_cpus=4):
+
+def build_vocab_from_text_file(file_path: str, tokenizer: Optional[Callable] = None, min_freq: int = 1, num_cpus: int = 4) -> Vocab:
     r"""Create a `Vocab` object from a raw text file.
-
     The `file_path` can contain any raw text. This function applies a generic JITed tokenizer in
     parallel to the text.
 
     Args:
-        file_object (FileObject): a file object to read data from.
-        jited_tokenizer (ScriptModule): a tokenizer that has been JITed using `torch.jit.script`
+        file_object: A file object to read data from.
+        tokenizer: A python callable to split input sentence into tokens. It can also be a Jited Module. By default, the function will do tokenization based on python split() function.
         min_freq: The minimum frequency needed to include a token in the vocabulary.
-            Values less than 1 will be set to 1. Default: 1.
-        num_cpus (int): the number of cpus to use when loading the vectors from file. Default: 4.
+        num_cpus: the number of cpus to use when loading the vectors from file.
+    Returns:
+        torchtext.vocab.Vocab: a `Vocab` object.
+    Examples:
+        >>> from torchtext.vocab import build_vocab_from_text_file
+        >>> v = build_vocab_from_text_file('vocab.txt')
+    """
+
+    if not tokenizer:
+        def tokenizer(x):
+            return x.split()
+
+    if isinstance(tokenizer, torch.jit.ScriptModule) or isinstance(tokenizer, torch.jit.ScriptFunction):
+        vocab_obj = _build_vocab_from_text_file(file_path, min_freq, num_cpus, tokenizer)
+    else:
+        vocab_obj = _build_vocab_from_text_file_using_python_tokenizer(file_path, min_freq, num_cpus, tokenizer)
+    return Vocab(vocab_obj)
+
+
+def load_vocab_from_file(file_path: str, min_freq: int = 1, num_cpus: int = 4) -> Vocab:
+    r"""Create a `Vocab` object from a text file.
+    The `file_path` should contain tokens separated by new lines.
+    Format for txt file:
+
+        token1
+        token2
+        ...
+        token_n
+
+    Args:
+        file_object: A file like object to read data from.
+        min_freq: The minimum frequency needed to include a token in the vocabulary.
+        num_cpus: the number of cpus to use when loading the vectors from file.
 
     Returns:
         torchtext.vocab.Vocab: a `Vocab` object.
 
     Examples:
-        >>> from torchtext.vocab import build_vocab_from_text_file
-        >>> from torchtext.experimental.transforms import basic_english_normalize
-        >>> tokenizer = basic_english_normalize()
-        >>> tokenizer = basic_english_normalize()
-        >>> jit_tokenizer = torch.jit.script(tokenizer)
-        >>> v = build_vocab_from_text_file('vocab.txt', jit_tokenizer)
+        >>> from torchtext.vocab import load_vocab_from_file
+        >>> v = load_vocab_from_file('vocab.txt')
     """
-    vocab_obj = _build_vocab_from_text_file(file_path, min_freq, num_cpus, jited_tokenizer)
+
+    vocab_obj = _load_vocab_from_file(file_path, min_freq, num_cpus)
     return Vocab(vocab_obj)
