+++ torchtext/vocab.py
@@ -8,138 +8,21 @@ import torch.nn as nn
 from urllib.request import urlretrieve
 from tqdm import tqdm
 import tarfile
-from typing import Optional, List, Dict, Callable, Iterable
+from typing import Dict, List, Optional, Iterable
+from collections import Counter, OrderedDict
 from torchtext._torchtext import (
     Vocab as VocabPybind,
-    _load_vocab_from_file,
-    _build_vocab_from_text_file_using_python_tokenizer
 )
-from collections import Counter, OrderedDict
 from .utils import reporthook
 
 logger = logging.getLogger(__name__)
 
 __all__ = [
-    'build_vocab_from_text_file',
-    'load_vocab_from_file',
     'build_vocab_from_iterator',
     'vocab',
 ]
+
 logger = logging.getLogger(__name__)
-
-
-def build_vocab_from_text_file(file_path: str, tokenizer: Optional[Callable] = None, min_freq: int = 1, num_cpus: int = 4):
-    r"""Create a `Vocab` object from a raw text file.
-
-    The `file_path` can contain any raw text. This function applies a generic JITed tokenizer in
-    parallel to the text.
-
-    Args:
-        file_object: A file object to read data from.
-        tokenizer: A python callable to split input sentence into tokens. By default,
-            the function will do tokenization based on python split() function.
-        min_freq: The minimum frequency needed to include a token in the vocabulary.
-        num_cpus: the number of cpus to use when loading the vectors from file.
-
-    Returns:
-        torchtext.vocab.Vocab: a `Vocab` object.
-
-    Examples:
-        >>> from torchtext.vocab import build_vocab_from_text_file
-        >>> v = build_vocab_from_text_file('vocab.txt')
-    """
-    if not tokenizer:
-        def tokenizer(x):
-            return x.split()
-
-    vocab_obj = _build_vocab_from_text_file_using_python_tokenizer(file_path, min_freq, num_cpus, tokenizer)
-    return Vocab(vocab_obj)
-
-
-def load_vocab_from_file(file_path: str, min_freq: int = 1, num_cpus: int = 4):
-    r"""Create a `Vocab` object from a text file.
-    The `file_path` should contain tokens separated by new lines.
-    Format for txt file:
-
-        token1
-        token2
-        ...
-        token_n
-
-    Args:
-        file_object: A file like object to read data from.
-        min_freq: The minimum frequency needed to include a token in the vocabulary.
-        num_cpus: the number of cpus to use when loading the vectors from file.
-
-    Returns:
-        torchtext.vocab.Vocab: a `Vocab` object.
-
-    Examples:
-        >>> from torchtext.vocab import load_vocab_from_file
-        >>> v = load_vocab_from_file('vocab.txt')
-    """
-
-    vocab_obj = _load_vocab_from_file(file_path, min_freq, num_cpus)
-    return Vocab(vocab_obj)
-
-
-def build_vocab_from_iterator(iterator: Iterable, min_freq: int = 1):
-    """
-    Build a Vocab from an iterator.
-
-    Args:
-        iterator: Iterator used to build Vocab. Must yield list or iterator of tokens.
-        min_freq: The minimum frequency needed to include a token in the vocabulary.
-    """
-
-    counter = Counter()
-    for tokens in iterator:
-        counter.update(tokens)
-    sorted_by_freq_tuples = sorted(counter.items(), key=lambda x: x[1], reverse=True)
-    ordered_dict = OrderedDict(sorted_by_freq_tuples)
-    word_vocab = vocab(ordered_dict, min_freq=min_freq)
-    return word_vocab
-
-
-def vocab(ordered_dict: Dict, min_freq: int = 1):
-    r"""Factory method for creating a vocab object which maps tokens to indices.
-
-    Note that the ordering in which key value pairs were inserted in the `ordered_dict` will be respected when building the vocab.
-    Therefore if sorting by token frequency is important to the user, the `ordered_dict` should be created in a way to reflect this.
-
-    Args:
-        ordered_dict: Ordered Dictionary mapping tokens to their corresponding occurance frequencies.
-        min_freq: The minimum frequency needed to include a token in the vocabulary.
-
-    Examples:
-        >>> from torchtext.vocab import vocab
-        >>> from collections import Counter, OrderedDict
-        >>> counter = Counter(["a", "a", "b", "b", "b"])
-        >>> sorted_by_freq_tuples = sorted(counter.items(), key=lambda x: x[1], reverse=True)
-        >>> ordered_dict = OrderedDict(sorted_by_freq_tuples)
-        >>> v1 = vocab(ordered_dict)
-        >>> print(v1['a']) #prints 1
-        >>> print(v1['out of vocab']) #raise RuntimeError since default index is not set
-        >>> tokens = ['e', 'd', 'c', 'b', 'a']
-        >>> v2 = vocab(OrderedDict([(token, 1) for token in tokens]))
-        >>> #adding <unk> token and default index
-        >>> unk_token = '<unk>'
-        >>> default_index = -1
-        >>> if unk_token not in v2: v2.insert_token(unk_token, 0)
-        >>> v2.set_default_index(default_index)
-        >>> print(v2['<unk>']) #prints 0
-        >>> print(v2['out of vocab']) #prints -1
-        >>> #make default index same as index of unk_token
-        >>> v2.set_default_index(v2[unk_token])
-        >>> v2['out of vocab'] is v2[unk_token] #prints True
-    """
-
-    tokens = []
-    for token, freq in ordered_dict.items():
-        if freq >= min_freq:
-            tokens.append(token)
-
-    return Vocab(VocabPybind(tokens, None))
 
 
 class Vocab(nn.Module):
@@ -311,6 +194,81 @@ class Vocab(nn.Module):
             cpp_vocab = torch.classes.torchtext.Vocab(self.vocab.itos_, self.vocab.default_index_)
             return Vocab(cpp_vocab)
         return self
+
+
+def vocab(ordered_dict: Dict, min_freq: int = 1) -> Vocab:
+    r"""Factory method for creating a vocab object which maps tokens to indices.
+
+    Note that the ordering in which key value pairs were inserted in the `ordered_dict` will be respected when building the vocab.
+    Therefore if sorting by token frequency is important to the user, the `ordered_dict` should be created in a way to reflect this.
+
+    Args:
+        ordered_dict: Ordered Dictionary mapping tokens to their corresponding occurance frequencies.
+        min_freq: The minimum frequency needed to include a token in the vocabulary.
+
+    Returns:
+        torchtext.vocab.Vocab: A `Vocab` object
+
+    Examples:
+        >>> from torchtext.vocab import vocab
+        >>> from collections import Counter, OrderedDict
+        >>> counter = Counter(["a", "a", "b", "b", "b"])
+        >>> sorted_by_freq_tuples = sorted(counter.items(), key=lambda x: x[1], reverse=True)
+        >>> ordered_dict = OrderedDict(sorted_by_freq_tuples)
+        >>> v1 = vocab(ordered_dict)
+        >>> print(v1['a']) #prints 1
+        >>> print(v1['out of vocab']) #raise RuntimeError since default index is not set
+        >>> tokens = ['e', 'd', 'c', 'b', 'a']
+        >>> v2 = vocab(OrderedDict([(token, 1) for token in tokens]))
+        >>> #adding <unk> token and default index
+        >>> unk_token = '<unk>'
+        >>> default_index = -1
+        >>> if unk_token not in v2: v2.insert_token(unk_token, 0)
+        >>> v2.set_default_index(default_index)
+        >>> print(v2['<unk>']) #prints 0
+        >>> print(v2['out of vocab']) #prints -1
+        >>> #make default index same as index of unk_token
+        >>> v2.set_default_index(v2[unk_token])
+        >>> v2['out of vocab'] is v2[unk_token] #prints True
+    """
+
+    tokens = []
+    for token, freq in ordered_dict.items():
+        if freq >= min_freq:
+            tokens.append(token)
+
+    return Vocab(VocabPybind(tokens, None))
+
+
+def build_vocab_from_iterator(iterator: Iterable, min_freq: int = 1) -> Vocab:
+    """
+    Build a Vocab from an iterator.
+
+    Args:
+        iterator: Iterator used to build Vocab. Must yield list or iterator of tokens.
+        min_freq: The minimum frequency needed to include a token in the vocabulary.
+
+    Returns:
+        torchtext.vocab.Vocab: A `Vocab` object
+
+    Examples:
+        >>> #generating vocab from text file
+        >>> import io
+        >>> from torchtext.vocab import build_vocab_from_iterator
+        >>> def yield_tokens_batch(file_path):
+        >>>     with io.open(file_path, encoding = 'utf-8') as f:
+        >>>         for line in f:
+        >>>             yield line.strip().split()
+        >>> vocab = build_vocab_from_iterator(yield_tokens_batch(file_path))
+    """
+
+    counter = Counter()
+    for tokens in iterator:
+        counter.update(tokens)
+    sorted_by_freq_tuples = sorted(counter.items(), key=lambda x: x[1], reverse=True)
+    ordered_dict = OrderedDict(sorted_by_freq_tuples)
+    word_vocab = vocab(ordered_dict, min_freq=min_freq)
+    return word_vocab
 
 
 def _infer_shape(f):
